chore: give Blocks "components"; not "templates".
@dwalkr pointed out that passing in the form templates to Blocks was
violating the single-responsibilitty principle. Templates are
responsible for defining how a block's data is ot be edited. The
Blocks component is responsible for rendering. The previous
design added a responsibility to the block's form template and also
gave the Blocks component access to more information then necessary.

This design keeps responsibilities a bit more isolated. Also note that
it greatly simplifies the Blocks component, cuts the amount of
tests it needs in half, and removes a dependency.