fix(show file): sanitize path URLs for windows.
Possible fix for several reports

There are three pieces to this bug.

First, on windows path's can contain forward slashes, or
double back slashes. For example

    data/author.json

or

    data\\author.json

Second, Node's `path.join` returns the second on Windows.

Third, the `git show` command behaves oddly with the former:

    git show HEAD:data/author.json

Returns the contents of the file

    git show HEAD:data/dne.json

Throws an error, and finally

    git show HEAD:data\\author.json

returns an empty string.

This very well may be a bug in Git. I'm not sure.