refactor(sidebar state): moved onto TinaCMS
This change reifies the sidebar state and places
it on the TinaCMS class as a property. It also
removes the need for the `SidebarContext`
in the React app, instead relying on the
pre-existinig CMSContext.

Here is how it would be used.

Checking sidebar state:

    cms.sidebar.isOpen

Changing sidebar state:

    cms.sidebar.isOpen = true

Subscribing Option 1

    import { useCMS, useSubscribable } from "tinacms"
    const cms = useCMS()

    useSubscribable(cms.sidebar)

Subscribing Option 2:

    const sidebar = useSidebar()