feat(@tinacms/form-builder,@tinacms/react-forms): Combines FormView and FormBuilder
In an effort to both simplify and address inconsistencies in forms, we've decided to merge `FormView` (which was exclusively used by the Sidebar and Screen Forms) and `FormBuilder` (which was used by `Content Creator` and other modal forms).

Initially, this should result in very few changes in userland.  Those will come later as we seek to change how `<FormActionMenu />` behaves and more.