chore: Resolves/ignores typescript issues related to Forms
Many of these appear to be pre-existing typescript issues that have appeared now that we've moved forward with `final-form` and `final-form-types`.  A few of them were hidden in the way we were building forms (especially inline forms), only exposed once I started cleaning up the code and simplifying the structure (`<FormLegacy />`).