fix(react-tinacms-editor,react-tinacms-inline): Fixes deps
Like the work on `react-tinacms-github` and `next-tinacms-github`, `@tinacms/form-builder` should no longer be a direct `dependency`, but instead a `peerDependency/devDependency`.  Having it a direct depdendency was causing issues downstream with (I believe) the use of `FormPortalProvider` which would have relied on a shared context.

Can confirm that I was able to using the `MarkdownFieldPlugin` once I made this change using `beta` releases of both packages and the GitHub Backend flow.