fix(@tinacms/toolkit): Removes onKeyPress from FormBuilder
A lingering difference between our `ContentCreator` (and Modal Forms) vs. Sidebar is that Modal Forms were coded to submit on `Enter`.  While handy, this meant that certain fields - namely `textarea` - would not function properly.

This behavior was carried over to `FormBuilder` as part of the merger, but is better left removed entirely.