fix(tinacms): Test if currentFields is an Array
This cropped up when adding a new NavItem to the Starter.  The error message:

`currentFields.find() is not a function.`

Debugging this, it appears that `currentFields` is not always a function of `fields` and can sometimes be a singular `field`.

The fix appears to be simply to test if `currentFields` is an Array, otherwise, the value of `currentFields` is already a singular `field`.