visually link Block data in sidebar and document
When hovering over the block list item in the sidebar,
a dotted line will appear around the corresponding
portion of the document that was generated
from that data.

This must be implemented on the consumer side
by adding a `data-tinafield` attribute to the DOM
and including the full field name as recognized by Tina.