2113 - Prevent nagivation away from an invalid form
One of our users pointed out an interesting way to circumvent validation on forms:

If you have a `group`, `group-list`, `blocks`, or multiple forms, you can use the "back" arrow in the form header to navigate away from an invalid form and use the Save/Reset buttons to commit those invalid changes.

The reason for this is because we're only using the currently rendered `finalForm`'s `invalid` flag to determine the state of those buttons rather than looking at all of the changes and validating them.

While a much bigger solution would probably exist in something like "Nested Forms," a much simpler solution is to simply prevent navigating away from an `invalid` form.