2122 - Adds `getDocumentFields()` query
Mainly for use with the upcoming Tina Admin, I've introduced a `getDocumentFields()` query which seeks to address the issue of retrieving the `fields` for a document before the document is created.

While `getDocumentFields()` has gone through a few iterations (and a couple of names), this version is the most... design agnostic, serving mostly as a `collections-templates-fields` tree that we can use where needed in the Tina Admin.

Very good chance it will change, grow, or become more specific as the product takes shape.