2156 - Adds Login/Logout support to TinaAdmin
Due to issues with the `EditStateContext`, `TinaAdmin` had to use the pure functions - `isEditing` and `setEditing` to modify the local storage key for `tina.isEditing`.

This PR adds a simple `LoginPage.tsx` along with a couple of new links to the `index.tsx` to either "Log out" or "Back to site".