2181 - Introduces RouteMapperPlugin and FormMetaPlugin
For a while now, `TinaAdmin` has mostly been a standalone wildcard route and has little interaction with TinaCMS inside the site.

This PR aims to provide ways for the two to interact - if only slightly.

On the `TinaAdmin` side of things, I've introduced a `RouteMapperPlugin` type that is set up inside the `cmsCallback`.  This provides the end-user with a `collection` and `document` and expects them to return the route to that document in the live site (or undefined, if there isn't one).

On the `TinaCMS` side of things, I've introduced the `FormMetaPlugin` which is an attempt to empower the user to add `meta` information to the `Form` instance.  In this case, `TinaAdmin` will place a link in the header (or where ever Scott B designs) that allows for linking back to that document in `TinaAdmin`.

This is a really barebones implementation that I expect us to iterate on.