2163 - Returns optional `collection` args for generic GraphQL queries
Credit to Jeff for discovering this quickly, but the upcoming changes to the GraphQL queries might break some existing uses of `updateDocument()` when they are introduced for `TinaAdmin`.

Before `createDocument` and `updateDocument` took a required argument `collection`, but more recently, we've discovered that argument is really necessary.

Originally, I removed the argument entirely from those queries, but because `TinaCloud` is not up-to-date with the latest `cli` changes, we've decided to bring the argument back (optional, instead of required).

This should ensure some degree of backwards compatability.