2258 - Swaps starter's old `admin` for the new one
From a conversation during office hours, we decided to rework `TinaAdmin` to provide the same login/logout functionality that the old `pages/admin/[...slug].tsx` and enable us to replace that file with the new `pages/admin/[...tina].tsx` that contains the import of `TinaAdmin`.

With this in place, a user can turn on and off `TinaAdmin`'s dashboard, content creator, etc. through the `tina-admin` flag without also having to do any file manipulation.

It also simplifies what the CLI needs to do dramatically.