2289 - Upgrades `TinaAdmin` to `react-router v6`
Aside from a few snags (removal of `Switch` in favor of `Routes`, `Route` taking a more open `element` prop instead of `children`, `NavLink` having `className` as a function), this was a fairly easy port to make.

I was trying a few different things, but I wanted to get this direct upgrade out first before I really started to experiment.