2335 - Prevents navigation into objects when parent form is invalid
Pointed out by a Tina User, while we prevented navigation away from a form when the form was invalid, we did _not_ prevent navigation into objects (`block`, `group`, `group-list`).

This could result in a frustrating situation where the user could not navigate back to the invalid form to fix its invalid state.

This is a more comprehensive fix than the adjustment I made in #2115.

Closes #2335