2374 - Incorrect messaging on admin when in local-mode
When in `local-mode`, `TinaAdmin` was using old messaging about logging into "Tina Cloud".

Ideally, we'd have a way to know if we're in `localMode` or not and alter the messaging on the Login page, but unfortunately, our indicator of `localMode` is attached to a `cms` instance that doesn't exist before you've logged in.

As a quick fix, I removed the message entirely and altered the `Log In` button's label to speak more about `edit-mode` - which exists in either scenario.