2357 - Adds TinaAdminAPI and consolidates Admin types
My goal was to pull out the direct GraphQL calls into a centralized class and consolidate the types that TinaAdmin uses. While simple at this point, my intent is to use the same TinaAdminApi in both TinaAdmin as well as the new Sidebar so that both places are using the same queries.