2512 - Fixes support for collection.templates in graphql
Looks like, as part of the datalayer work, the `keepTemplateKey` boolean was replaced by a hardcoded `false` that essentially removed support for `collection.templates` shapes in schema.

None of our tests picked up on this because our tests have their own `get/put` commands - this only happened in the `filesystem` and `github` stores.

The solution is to keep track of that flag and pass it around.