2530 - Updated relativePath field on Create Page
For a non-developer, the "Relative Path" field can be confusing.  In reality, this is just the "Filename" where the Document will reside (against the collection's `path` and `format`).

I've taken the old "Name" field from the `ContentCreator` and ported it over to the Collection Create Page, removing the need to specify the file extension and adding some simple validation around the field.