2540 - Uses onbeforeunload events to prevent navigation away from dirty forms
This is more of a stop gap solution for something much more complex.

As we continue to blur the lines between the Sidebar Experience and the Admin Experience, a user may unknowingly seek to navigate to a Collection List page while in the middle of editing a page.  Right now, the user would navigate to the Collection List page and lose the entirety of their changes in an "dirty" form.

This solution uses `<FormSpy />` to emit a single `onbeforeunload` event when the form is "dirty" (!pristine) and removes that event when either the form is not "dirty" _or_ has been unmounted.

Navigation _within_ the app or _within_ the admin is, as of right now, not prevented.