2596 - Changes how RouteMapping Plugin is used to allow for dynamic import
Scott found an interesting race condition when trying to dynamically import `RouteMappingPlugin` via the cmsCallback.

This resulted in the overriden links to only be present on subsequent loads of the List page rather than the first (because the plugin had not been attached yet).

I've changed the code so that RouteMapping is called when needed (when the link is clicked, for example).