2548 - Adds activity indicator to main panel of Admin
While locally, the Admin is snappy, in production or on Tina Cloud, there is some delay in loading a Collection List or Document.

Before, there was no way to tell that the request had even been made - resulting in potentially repeat clicking (which would further increase the delay) or a general lack of confidence that the Admin was even working.  This PR adds a simple acitivity indicator through the Admin for _most_ of the requests.