2600 - Handle errors better in the CMS
Right now, the CMS was built for "happy path" and does not particularly hold up well when it encounters an error from the GraphQL server.

While an enterprising developer can poke through the network log and find the issue, I would like to surface _something_ to the user if at least to get the ball rolling on what might've went wrong.

In the perfect world, errors that make it to the CMS would be caught earlier by the CLI or the GraphQL server so the chance for an error is rare.

Closes #2600