2635 - Minor change to the timeout for success messaging
Scott and I added the 30 second delay on **error** messaging because we felt the default timer for `alerts` is a bit short for something important like that.

For standard messages, like **success** toasts, we can stick to the default (we don't want a lingering message to be misleading if the user is navigating multiple Documents).

Related to #2635