2707 - Allows Documents to be created with subfolders in the filename
Brought up in a Discord discussion, we currently do not support Documents nested within subfolders within the CMS.  The documents can be queried, but creating and updating via the CMS fails.

This PR allows for a `/` to be used in the `filename` field and accounts for nested documents by using the document's `breadcrumbs` (which also needed to be fixed).