Adds `useRelativeMedia` to local GraphQL
Part of the Repo-based Media push, this adds an extra `config` to `resolve()` that contains (for now) just `useRelativeMedia: true`.

Locally, this value will always be `true`.

In the future, a companion PR to `tina-cloud` will include a slightly different `config`:
```
{ useRelativeMedia: false, clientId: app.client_id }
```

which will be used to determine the location of the file on `assets.tinajs.io`.