2935 - Fixes subitem links to use breadcrumbs
Because the CMS supports sub-directories, we need to account for the potential existence of those in any navigation.

The sublinks for items in the CMS were only using `filename` (which is only the `basename`).
In other places, we use `breadcrumbs` which is an array of all directories leading up to the `basename`).

Closes #2935