Optionally find Tina files in code repo with FilesystemBridge
This updates the GitHubBridge to extend the FilesystemBridge and optionally fallback to the latter when fetching Tina files. This provides a path to support self-hosting with separate content repos without needing those separate content repos to track all the Tina files from the code repository. We can instead specifically opt to use the Tina files from the local filesystem in the code repository.