fix: update branch creation modal text for clarity and protection notice (#5580)
This pull request includes changes to the `CreateBranchModel` component
in `packages/tinacms/src/toolkit/form-builder/form-builder.tsx`. The
changes focus on improving the user interface and user experience when
creating a new branch.

User interface and experience improvements:

* Removed the `currentBranch` variable as it was not being used in the
component.
* Updated the modal content to provide clearer instructions and a more
engaging message for users. The text now includes a bold header and a
more descriptive explanation of the branch creation process.
* Changed the placeholder text for the branch name input field to give
users a clearer example of the expected input format.

![CleanShot 2025-03-18 at 17 16
08](https://github.com/user-attachments/assets/6e8026a6-0bf1-49bc-aa13-020607b30052)