Rev-package-versions (#5589)
I Accidentally published packages by tagging the 'Version Packages' PR,
so I deprecated the published versions.

This pull request includes a series of version bumps across multiple
`package.json` files and a modification to a GitHub workflow. The
changes ensure that dependencies are up to date and refine the workflow
conditions.

### Workflow Update:
*
[`.github/workflows/main.yml`](diffhunk://#diff-7829468e86c1cc5d5133195b5cb48e1ff6c75e3e9203777f6b2e379d9e4882b3L90-R91):
Added a condition to exclude pull requests with the title 'Version
Packages' from triggering the workflow.