⚡ Performance - Improve media load performance (#5995)
When loading assets in TinaCMS we do a redundant call to the user GET
endpoint just to check authentication.

As we already have logic in place to handle Unauthorized errors so when
we do the call to get the assets, this authentication is handled.

Therefore by removing this extra call, we can speed up the media
loading.