GraphQL Test - Top-level templates (#6012)
This change adds tests for the situation where the top-level entry in a
collection defines templates.

These tests ensure that the 'unionData' lookup type branch in the field
resolver is exercised -
https://github.com/tinacms/tinacms/blob/304f1e2725f12e3ad12d3a44d2aa73ce68bab3fc/packages/%40tinacms/graphql/src/resolve.ts#L376

This PR includes two commented-out blocks for tests that demonstrate an
existing bug in the field resolver. As the lookup generation fails to
record the name of the `create{{ COLLECTION }}` field as a 'create'
field, the field reducer fails to process those fields correctly despite
being permitted to by the GraphQL schema.