GraphQL Test - Renaming documents (#6020)
This change adds a test for renaming a document, as it is done through
the TinaCMS admin interface.

Please note the following:

- The `update{{ COLLECTION-NAME}}` mutation does not permit renaming, so
no test for it is included.
- It is possible to included updated content when renaming a document.
However, that updated content is ignored by the field resolution
process, so no test for it is included.
- Due to the increase in the number of snapshots in the test, all
snapshots have been moved to a dedicated 'expected-snapshots' directory.
- To prevent deleting content files during the test, the
`MemoryCaptureBridge` is extended to intercept delete requests. That
way, tests can be run more than once without needing to restore content
files.