Remove redundant icons in modals (#6107)
This PR is a fix to remove the redundant icon in the editorial workflow
modal as we now have the icon in all modal headers.

<!--

Thank you for contributing to TinaCMS!

Several things must happen for your PR to be merged:

1. It must successfully build, test, and lint your branch
1. It must pass the dangerjs script
2. It must be up-to-date with master
3. It must be approved by at least 1 tinacms core member
4. It must have conventional-commits that clearly explain what has been
changed


Small is beautiful! 

PRs should be small. They should be made up of many small commits, with
clear
commit messages explaining _what_ has been changed and _why_. The tests
should
be short and specific, with single assertions. 

-->