🗒️Date Picker - Overhaul using SchadCN component (#6174)
### Description

This PR is a rewrite of the existing date picker. The new date picker
uses this [Schadcn
component](https://shadcnui-expansions.typeart.cc/docs/datetime-picker)
as a base, but I've configured it to work with contextual editing, match
our existing design system, and work with the existing configuration
options we had for the legacy date picker.

**Note: I had to fix some z indexing issues that were causing the
calendar to overlap with the sidebar when the user scrolls**

### Screenshots

![CleanShot 2025-11-26 at 10 17
18](https://github.com/user-attachments/assets/f7314fad-4bc3-4c4b-9910-5089b1f2b8fc)
**Figure: Typical editing flow with date picker**

<br />

![CleanShot 2025-11-26 at 10 15
33](https://github.com/user-attachments/assets/5381304f-dc85-4181-a4eb-c611eac5e6b3)
**Figure: Date picker opening using contextual editing**