Resolver - Refactor to improve clarity (#6084)
CC @brookjeynes-ssw 

The main purpose of this change is to improve the readability of the
GraphQL resolver, thereby making it easier to resolve bugs and fix
inefficiencies in TinaCMS' GraphQL query handling.

The PR attempts to do the following:

- Clarify the logic within the field resolver defined in `resolve.ts` so
that the code-path executed on the `Resolver` object is obvious.
- To support this object, the functionality of the
`Resolver.resolveDocument` method is split out into new methods:
`resolveAddPendingDocument`, `resolveRetrievedDocument`,
`resolveCreateFolder`, `resolveCreateDocument`, `resolveUpdateDocument`
and `resolveDeleteDocument`.
- Previously existing public methods of the `Resolver` that are no
longer required are marked as deprecated in preparation for removal at
the next major version. They are not removed now as they are currently
part of the public interface of the package, though they shouldn't be
used outside of the `@tinacms/graphql` package.
- Improves type and shape checking throughout the resolution process, to
make it clearer what types are expected at each point in the process.
- Simplifies the logic of the `getOptimizedQuery` query.
- Removes leading underscores from parameter names where the parameters
are used.
- Replaces occurrences of `@ts-ignore` with `@ts-expect-error`, to catch
future situations where suppression is no longer required for a line due
to either improvements in TypeScript or changes to the line.

This change aims to maintain the existing behaviour of the resolver,
with the exception of one bug affecting "create{{ COLLECTION-NAME }}"
queries where the top-level entry in a collection does not have
templates. This is because additional logic would need to be added to
reinstate the bug. A test is added to confirm the correction and prevent
regression. Note that the bug still exists for "create{{ COLLECTION-NAME
}}" queries where the top-level entry does have templates.

## Issues resolved

- https://github.com/tinacms/tinacms/issues/5925
- https://github.com/tinacms/tinacms/issues/5926

## Future directions for consideration

- The `Resolver` could be reduced to two public methods that could be
called from `resolve`: `typeResolver` and `fieldResolver`. That way, all
other methods could be made private to minimise the public interface of
the class. Note that this would require that the functions in
`auth-fields.ts` are folded into `Resolver.
- Further bug fixes and improvements to the Resolver will require
changes to the generated database Lookup, but any changes to the
generated Lookup will affect the schema checks conducted during `tinacms
build`.

## Testing

I have not been able to test this change as thoroughly as I would like.
The reviewer must test this change with sophisticated code bases such as
https://github.com/tinacms/tina.io .

---------

Co-authored-by: Pat Stuart [SSW] <71543754+Pat-Stuart@users.noreply.github.com>